services/requirements.txt
google-genai>=0.7.0
python-dotenv>=1.0.0
firebase-admin>=6.4.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
logging